refactor(templates): clean all the template with host bindings
There are a lot of useless div containers just for properties binding. That's unnecessary as the components themselves are already containers. Of course, there was some benefits for styling and theming. But that could be later styling redesigning or just extracting.

NOTE: As a performant tree view, we should keep our tags as necessary as possible.

And add animation for tree-node-children.